firmware cleanup
