+++ Firmwares/SerialDev/serial_led.h
@@ -1,0 +1,205 @@
+// #include "FastLED.h"
+// #include "globals.h"
+
+#define FIRMWARE_VER 3
+
+const uint8_t deviceID = 0;
+
+namespace CMDTYPE
+{
+    enum CMDTYPE
+    {
+        SETUP_DATA = 1,
+        PIXEL_DATA = 2,
+        BRIGHTNESS = 3,
+        GETID      = 4,
+        SETID      = 5,
+        GETVER     = 6,
+        GETBTNS    = 8
+    };
+}
+
+namespace RETURN_CODES
+{
+    enum RETURN_CODES
+    {
+        SUCCESS = 255,
+        REBOOT = 42,
+        ERROR = 0,
+        ERROR_SIZE = 1,
+        ERROR_UNSUPPORTED = 2,
+        ERROR_PIXEL_COUNT = 3,
+        ERROR_BAD_CMD = 4,
+    };
+}
+
+typedef struct __attribute__((__packed__))
+{
+	uint8_t type;
+	uint16_t pixelCount;
+	uint8_t spiSpeed;
+} config_t;
+
+uint16_t numLEDs = NUM_LEDS;
+uint8_t bytesPerPixel = 3;
+
+CRGB leds[NUM_LEDS];
+CLEDController * pLed = NULL;
+
+inline void setup_leds() {
+//   FastLED.addLeds<LED_TYPE,DATA_PIN,CLK_PIN,COLOR_ORDER>(leds, NUM_LEDS).setCorrection(TypicalLEDStrip);
+
+    pLed = new APA102Controller<DATA_PIN, CLK_PIN, COLOR_ORDER>();
+    FastLED.addLeds(pLed, leds, NUM_LEDS);
+    pLed->scale(DEFAULT_BRIGHTNESS);
+
+  // set master brightness control
+//   FastLED.setBrightness(BRIGHTNESS);
+
+  FastLED.clear();
+  FastLED.show();
+}
+
+#define EMPTYMAX 100
+inline void getData()
+{
+    static char cmd = 0;
+    static uint16_t size = 0;
+    static uint16_t count = 0;
+    static uint8_t emptyCount = 0;
+    static size_t c = 0;
+    static uint16_t packSize = numLEDs * bytesPerPixel;
+
+    if (Serial.available())
+    {
+        cmd = Serial.read();
+        size = 0;
+        Serial.readBytes((char*)&size, 2);
+
+        if (cmd == CMDTYPE::PIXEL_DATA)
+        {
+            count = 0;
+            emptyCount = 0;
+
+            if (size == packSize)
+            {
+                while (count < packSize)
+                {
+                    c = Serial.readBytes(((char*)&leds) + count, packSize - count);
+                    if (c == 0)
+                    {
+                        emptyCount++;
+                        if(emptyCount > EMPTYMAX) break;
+                    }
+                    else
+                    {
+                        emptyCount = 0;
+                    }
+
+                    count += c;
+                }
+            }
+
+            uint8_t resp = RETURN_CODES::SUCCESS;
+            if (count == packSize)
+            {
+                noInterrupts();
+                FastLED.show();
+                interrupts();
+            }
+            else
+                resp = RETURN_CODES::ERROR_SIZE;
+
+            Serial.write(resp);
+        }
+        else if(cmd == CMDTYPE::GETBTNS)
+        {
+            Serial.write(RETURN_CODES::SUCCESS);
+            noInterrupts();
+            Serial.write((uint8_t*)btns, 32);
+            interrupts();
+        }
+        else if(cmd == CMDTYPE::GETID)
+        {
+            #ifdef USE_EEPROM
+                Serial.write(EEPROM.read(16));
+            #else
+                Serial.write(deviceID);
+            #endif
+        }
+        else if(cmd == CMDTYPE::SETID)
+        {
+            if(size != 1)
+            {
+                Serial.write(RETURN_CODES::ERROR_SIZE);
+            }
+            else
+            {
+                uint8_t id = Serial.read();
+                #ifdef USE_EEPROM
+                    EEPROM.write(16, id);
+                #endif
+                Serial.write(RETURN_CODES::SUCCESS);
+            }
+        }
+        else if (cmd == CMDTYPE::SETUP_DATA)
+        {
+            // for(int i=0; i<size; i++) Serial.read();
+
+            uint8_t result = RETURN_CODES::SUCCESS;
+            config_t temp;
+
+            if (size != sizeof(config_t))
+            {
+                result = RETURN_CODES::ERROR_SIZE;
+            }
+            else
+            {
+                size_t read = Serial.readBytes((char*)&temp, sizeof(config_t));
+                if (read != size)
+                {
+                    result = RETURN_CODES::ERROR_SIZE;
+                }
+                else
+                {
+                    if(temp.pixelCount / bytesPerPixel != NUM_LEDS)
+                        result = RETURN_CODES::ERROR_PIXEL_COUNT;
+                }
+            }
+
+            Serial.write(result);
+        }
+        else if (cmd == CMDTYPE::BRIGHTNESS)
+        {
+            uint8_t result = RETURN_CODES::SUCCESS;
+            if (size != 1)
+                result = RETURN_CODES::ERROR_SIZE;
+            else
+            {
+                uint8_t brightness = 255;
+                size_t read = Serial.readBytes((char*)&brightness, 1);
+                if (read != size)
+                    result = RETURN_CODES::ERROR_SIZE;
+                else
+                {
+                    // FastLED.setBrightness(brightness);
+                    pLed->scale(brightness);
+                }
+            }
+
+            Serial.write(result);
+        }
+        else if (cmd == CMDTYPE::GETVER)
+        {
+            Serial.write(RETURN_CODES::SUCCESS);
+            Serial.write(FIRMWARE_VER);
+        }
+        else
+        {
+            Serial.write(RETURN_CODES::ERROR_BAD_CMD);
+        }
+
+
+        Serial.flush();
+    }
+}
